Fixing win x64 build with vs2015

+++ test/core/transport/connectivity_state_test.c
@@ -39,7 +39,7 @@
 
 #include "test/core/util/test_config.h"
 
-#define THE_ARG ((void *)0xcafebabe)
+#define THE_ARG ((void *)(size_t)0xcafebabe)
 
 int g_counter;
